% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSignature.R
\name{addSignature}
\alias{addSignature}
\title{Add new signature(s) to an existing set of mutation signatures}
\usage{
addSignature(sigmatrix2, sigmatrix1)
}
\arguments{
\item{sigmatrix2}{An object of class \emph{mutsig} describing the new signature.}

\item{sigmatrix1}{An object of class \emph{mutsig} describing the existing set of signatures.}
}
\value{
An object of \emph{mutsig} class.
}
\description{
Add new signature(s) to an existing set of mutation signatures, and  return the updated signature set.
}
\details{
Addition of one or more new mutation signature(s) to an existing set of mutation signatures upon which the updated signature set is returned.
}
\examples{
newSigMatrix=addSignature(sigmatrix2=signatures.cosmic[c('Signature.10','Signature.11'),], 
sigmatrix1=signatures.cosmic[c('Signature.1','Signature.2'),])
}
\seealso{
\code{\link{deleteSignature}}, \code{\link{renameSignature}} and \code{\link{mergeSignature}}.
}
\keyword{Add}
\keyword{a}
\keyword{mutation}
\keyword{of}
\keyword{set}
\keyword{signatures}
\keyword{signatures.}
\keyword{to}
