% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseControlSig.R
\name{caseControlSig}
\alias{caseControlSig}
\title{Identify signatures with significantly higher mutation burden in cases over controls.}
\usage{
caseControlSig(contextfreq.cases, contextfreq.controls, signatures.ref=signatures.cosmic, threshold=0.05, adjust="fdr")
}
\arguments{
\item{contextfreq.cases}{A data frame of class \emph{contextfreq} containing mutation frequency in tri-nucleotide contexts in case samples}

\item{contextfreq.controls}{A data frame of class \emph{contextfreq} containing mutation frequency in tri-nucleotide contexts in control samples}

\item{signatures.ref}{An object of class \emph{mutsig} comprising the set of signatures. \emph{Default: 'signatures.cosmic'}}

\item{threshold}{Threshold for uncorrected percentile score. \emph{Default: 0.05}}

\item{adjust}{Method for p-value correction for multiple testing. Options are as provided in the function \emph{p.adjust}. The default method is the \emph{"FDR"} method.}
}
\value{
An object providing the following information:
\enumerate{ 
  \item \strong{n.case}:             Number of case samples. 
  \item \strong{n.control}:		      Number of control samples. 
  \item \strong{case.weights}:		    A data frame containing estimated weights of known mutation signatures in the case samples. 
  \item \strong{control.weights}:		A data frame containing estimated weights of known mutation signatures in the control samples.
  \item \strong{p.value}:		        Indicates statistical significance of higher mutation burden of signatures in cases over controls.
  \item \strong{adjust.p.value:}		  P-values adjusted for multiple testing correction.
  }
}
\description{
Identify signatures with significantly higher mutation burden in case samples over control samples.
}
\details{
Identifes signatures with significantly higher mutation burden in case samples campared to control samples using the Wilcoxon Rank sum test. This is suitable when the samples are homogeneous within respective groups, and signatures significantly over-represented in cases relative to the controls are of interest.
}
\examples{
load(file = "data/contextfreq.cases_test.rda")
load(file = "data/contextfreq.controls_test.rda")
signif_signatures_obj=caseControlSig(contextfreq.cases=contextfreq.cases_test, contextfreq.controls=contextfreq.controls_test, signatures.ref=signatures.cosmic, threshold=0.05, adjust="fdr")
}
\seealso{
\itemize{ 
\item \code{\link[deconstructSigs]{signatures.cosmic}}, \code{\link{confidenceSig}} for robust signatures and \code{\link{enrichSig}} for enriched signatures in individual case samples.
\item To generate \emph{contextfreq} object from  \emph{snv} dataframe use \code{\link{processSNV}} and \code{\link{vcfToSNV}}.
}
}
\keyword{enriched}
\keyword{signatures}
