% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSignature.R
\name{clusterSignature}
\alias{clusterSignature}
\title{Hierarchical clustering of mutational signature.}
\usage{
clusterSignatures(sigmatrix)
}
\arguments{
\item{sigmatrix}{An input matrix of mutational signatures.}
}
\value{
A hierarchical clustering object formed from the input matrix of mutational signatures. This can then be plotted to show how the mutational signatures in the input matrix are related to one another.
}
\description{
Performs a hierarchical clustering analysis of an input mutational signature matrix using the Euclidean distance metric.
}
\examples{
hclust_obj=clusterSignature(sigmatrix=signatures.cosmic) 
\\code{\\link{signaturePCA}} and \\code{\\link{signaturesHeatmap}}
}
\keyword{cluster}
\keyword{mutation}
\keyword{signatures.}
