% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceSig.R
\name{confidenceSig}
\alias{confidenceSig}
\title{Provides interval of uncertainty for estimated weights of mutation signatures}
\usage{
confidenceSig(contextfreq.sample, subsample=0.8, iter=1000 , signatures.ref=signatures.cosmic, lbound=0.1, ubound=0.9, replace=FALSE)
}
\arguments{
\item{contextfreq.sample}{A sample from the dataframe of class \emph{contextfreq} containing mutation frequency in trinucleotide contexts.}

\item{subsample}{Proportion of mutations included during each subsampling. Default: \emph{0.8}  (80 percent)}

\item{iter}{Number of iterations of subsampling. Default: \emph{1000}}

\item{signatures.ref}{An object of class \emph{mutsig} comprising the set of signatures. \emph{(signatures.nature2013 or signatures.cosmic or signatures.cosmic.2019 )}, Default: \emph{'signatures.cosmic'}}

\item{lbound}{Lower bound of the interval of uncertainty for estimated weights of the signatures Default: \emph{0.1} (10 percent)}

\item{ubound}{Upper bound of the interval of uncertainty for estimated weights of the signatures Default: \emph{0.9} (90 percent)}

\item{replace}{should sampling be with replacement? \emph{TRUE or FALSE}}
}
\value{
An object containing the following information:
\enumerate{ 
  \item \strong{observed.weights:} A data frame containing estimated weights of known mutation signatures based on all mutations in a sample.
  \item \strong{median.weights:} A data frame containing estimated median of the weights of known mutation signatures based on iteratively subsampled mutations.
  \item \strong{ubound.weights:} A data frame containing upper-bound values of the weights of known mutation signatures based on iteratively subsampled mutations.
  \item \strong{lbound.weights:} A data frame containing lower-bound values of the weights of known mutation signatures based on iteratively subsampled mutations.
  }
}
\description{
Provides an interval of uncertainty for estimated weights of known mutation signatures in the catalog of mutations from a set of samples.
}
\details{
Provides an interval of uncertainty for estimated weights of known mutation signatures in the catalog of mutations from a set of samples. First, based on the catalog of mutations in a sample, weights of the mutation signatures are estimated. Next, mutations are subsampled iteratively without replacement, each time estimating the weights of the mutation signatures. The intervals of uncertainty for weights of each mutation signatures are determined by aggregating observations from a given number of iterations.
}
\examples{
data(contextfreq.sample_test)
robust_sig_object=confidenceSig(contextfreq.sample=contextfreq.sample_test[1,], subsample=0.8, 
iter=50, signatures.ref=signatures.cosmic, lbound=0.1, ubound=0.9, replace=FALSE)
}
\seealso{
\itemize{ 
\item \code{\link[deconstructSigs]{signatures.cosmic}}, \code{\link{caseControlSig}} to identify signatures with significantly higher mutation burden in case samples over control samples and \code{\link{enrichSig}} for enriched signatures in individual case samples.  
 \item To generate \emph{contextfreq} object from \emph{snv} dataframe use \code{\link{processSNV}} and \code{\link{vcfToSNV}}.
 }
}
\keyword{estimated}
\keyword{for}
\keyword{mutation}
\keyword{of}
\keyword{signatures}
\keyword{uncertainty}
\keyword{weights}
