% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextSNV.R
\name{contextSNV}
\alias{contextSNV}
\title{Partition mutations according to their genomic contexts}
\usage{
contextSNV(snv, file = "empty", mode = "include")
}
\arguments{
\item{snv}{A dataframe having \emph{sample, chr, pos, ref, alt and/or  freq} as its columns.}

\item{file}{Path of the BED file, having genomic segments and their context information. This snv dataframe can be generated by \code{\link{vcfToSNV}}.}

\item{mode}{Default is set to \emph{include}: Select mutations lying inside the given genomic contexts (obtained from BED file). 
\emph{exclude}: will give mutations outside the given context.}
}
\value{
\emph{snv} dataframe with column names of \emph{sample, chr, pos, ref, alt, freq}. Where, sample name will be concatenated by the its genomic context. Example: if a mutation within a sample of name \emph{sample1}, lies in the context segment of label \emph{C1} given in the input \emph{bed file} , the sample name of mutation is modified to \emph{sample1_C1}. The mutations that don't lie in any context segment given in the \emph{BED} file are discarded.
}
\description{
Group mutations according to their occurrence within or outside user-annotated genomic segments.
}
\details{
When the \emph{include} mode is used, group mutations according to their occurrence in user-specifiedannotated genomic segments; such segments could represent specific genomic or epigenomic contexts (e.g.actively transcribed coding regions). Alternately, in the \emph{exclude} mode, mutations in undesired regions (e.g.black-listed genomic regions) can be removed.
}
\examples{
> BED_file=system.file("extdata", "context_testFile.bed", package = "MutSigTools", mustWork = TRUE)

> load(file = "data/snv_sample.rda")   # load 'snv' dataframe object

> context_snv=contextSNV(snv=snv_sample,BED_file, mode='include')

}
\seealso{
\code{\link{vcfToSNV}} to generate \emph{snv} dataframe, and \url{https://genome.ucsc.edu/FAQ/FAQformat.html} for BED file format.
}
