% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSignature.R
\name{deleteSignature}
\alias{deleteSignature}
\title{Delete selected signatures from an existing set of mutation signatures.}
\usage{
deleteSignature(sigmatrix, del_sig)
}
\arguments{
\item{sigmatrix}{An object of class \emph{mutsig} describing the existing set of signatures.}

\item{del_sig}{signature(s) to be removed.}
}
\value{
An object of \emph{mutsig} class with the deleted signatures.
}
\description{
Delete selected signatures from an existing set of mutation signatures.
}
\details{
Deletion of one or more mutation signatures from an existing set of mutation signatures upon which the updated signature set is returned.
}
\examples{
reducedSigMatrix=deleteSignature(sigmatrix=signatures.cosmic, c(2,5)) # delete signature 2 and 5.
}
\seealso{
\code{\link{addSignature}}, \code{\link{renameSignature}} and \code{\link{mergeSignature}}.
}
\keyword{delete}
\keyword{mutation}
\keyword{signatures.}
