% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSNV.R
\name{processSNV}
\alias{processSNV}
\title{Estimate mutation frequency in nucleotide motif contexts.}
\usage{
processSNV(snv, bsg = BSgenome.Hsapiens.UCSC.hg19::Hsapiens)
}
\arguments{
\item{snv}{A dataframe having \emph{sample, chr, pos, ref, alt and/or freq} as its columns. The \emph{snv} dataframe can be created by the \code{\link{vcfToSNV}} function.}

\item{bsg}{An object of class \emph{BSGenome}. Default: \emph{BSgenome.Hsapiens.UCSC.hg19::Hsapiens}}
}
\value{
A data frame of class \emph{contextfreq} containing mutation frequency in user-specified nucleotide contexts.
}
\description{
Estimate frequency of point mutations in user-specified nucleotide motif contexts (e.g. trincleotide motifs) for one or more input samples.
}
\details{
Estimation of mutation frequency in pre-determined, user-specified nucleotide motif contexts for input samples. The BSgenome object contains the reference genome information, and should be specified for alternative assembly of human genomes or nonhuman reference genomes. The output returns a data frame containing estimated mutation frequencies in the different genomic contexts, as provided in the \emph{contextfreq} object, for each input sample.
}
\examples{
load(file = "data/snv_sample.rda")
context.freq=processSNV(snv=snv_sample, bsg = BSgenome.Hsapiens.UCSC.hg19::Hsapiens)
}
\seealso{
\code{\link{vcfToSNV}} to generate \emph{snv} dataframe, \code{\link[BSgenome]{BSgenome}} and  \code{\link[BSgenome.Hsapiens.UCSC.hg19]{BSgenome.Hsapiens.UCSC.hg19}}.
}
\keyword{catalogue}
\keyword{mutation}
