% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_tSNE.R
\name{signature_tSNE}
\alias{signature_tSNE}
\title{Create tSNE plot of signatures.}
\usage{
signature_tSNE(sigmat, pngfile )
}
\arguments{
\item{sigmat}{An object of class \emph{mutsig} describing a set of signatures.}

\item{pngfile}{name of \emph{png} formatted image file to be created}
}
\value{
Create tSNE plot to show similarity among mutational signatures.
\enumerate{  
 \item \strong{*_tSNE.png} The tSNE plot of mutation signatures.
 }
}
\description{
Constructs tSNE plots of signatures.
}
\examples{
sigmat=signatures.cosmic
signature_tSNE(sigmat, pngfile="signature_tSNE")

}
\seealso{
\code{\link{signatureHeatmap}}, \code{\link{signaturePCA}}, \code{\link[deconstructSigs]{signatures.cosmic}}, \code{\link{deleteSignature}}, \code{\link{renameSignature}} and \code{\link{mergeSignature}}.
}
\keyword{mutation}
\keyword{signature}
\keyword{tSNE}
