% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureHeatmap.R
\name{signatureHeatmap}
\alias{signatureHeatmap}
\title{Create a heatmap of motif weights for signatures.}
\usage{
signatureHeatmap(sigmat, pngfile )
}
\arguments{
\item{sigmat}{An object of class \emph{mutsig} describing the existing set of signatures.}

\item{pngfile}{The name of \emph{png} formatted image file to be created.}
}
\value{
A \emph{png} formatted image file.
}
\description{
Construct a heatmap showing relative weights of tri-nucleotide motif for different mutation signatures.
}
\details{
Create a heatmap to show relative weights of trinucleotide motifs for different mutation signatures in a signature-set.
}
\examples{
sigmat=signatures.cosmic
signatureHeatmap(sigmat, pngfile='signatures_heatmap')
}
\seealso{
\code{\link{signaturePCA}}, \code{\link{clusterSignature}} , \code{\link{addSignature}} and \code{\link[deconstructSigs]{signatures.cosmic}}
}
\keyword{heatmap}
\keyword{mutation}
\keyword{signamture}
