% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaturePCA.R
\name{signaturePCA}
\alias{signaturePCA}
\title{Create three PCA plots for signatures.}
\usage{
signaturePCA(sigmat, pngfile )
}
\arguments{
\item{sigmat}{An object of class \emph{mutsig} describing a set of signatures.}

\item{pngfile}{name of \emph{png} formatted image file to be created}
}
\value{
Create three PCA plots with the following extensions to show difference among mutational signatures.
\enumerate{  
 \item \strong{*_Eigen.png} The eigenvalues of the PCA and the relative amount of the variation each eigenvalue.
 \item \strong{ *_var.png} The different single-nucleotide alteration components in terms of the two dimensions.
 \item \strong{*_PCA.png} The different signatures in terms of the 2 non-dimensional vectors derived by the PCA analysis and the cosine similarity between the estimate of each mutation signature using PCA and the actual mutation signature(in order to determine how accurately the 2 dimensions calculated by PCA represent the mutational signatures). 
 }
}
\description{
Constructs three PCA plots showing variations among the signatures in terms of the weights of the principal components for different mutation signatures, as well as the eigenvalues and cosine similarity among expected and actual signatures.
}
\examples{
sigmat=signatures.cosmic
signaturePCA(sigmat, pngfile="signaturePCA")

}
\seealso{
\code{\link{signatureHeatmap}}, \code{\link[deconstructSigs]{signatures.cosmic}}, \code{\link{deleteSignature}}, \code{\link{renameSignature}} and \code{\link{mergeSignature}}.
}
\keyword{PCA}
\keyword{mutation}
\keyword{signamture}
