% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfToSNV.R
\name{vcfToSNV}
\alias{vcfToSNV}
\title{Convert vcf file to snv dataframe}
\usage{
vcfToSNV(vcf, allelefreq = FALSE)
}
\arguments{
\item{vcf}{The path of a standard vcf file.}

\item{allelefreq}{A logical input, TRUE if mutation allele frequency is required in output object. Default : FALSE}
}
\value{
A snv dataframe having datframe with column names as sample, chr, pos, ref, alt, freq. Where, filename will be considered as sample name. \strong{Note}: vcf file name with no extension is the sample name.
}
\description{
Read vcf formatted file to catalog mutations in a data frame of class snv for downstream analysis. A snv dataframe consists of sample, chr, pos, ref, alt and/or freq columns.
}
\details{
Read vcf formatted file to catalog mutations in a data frame of class snv for downstream analysis foreach input sample.
}
\examples{
vcf_file=system.file("extdata", "test.vcf", package = "MutSigTools", mustWork = TRUE)

snv=vcfToSNV(vcf=vcf_file, allelefreq=TRUE)

head(snv)

}
\seealso{
\href{https://faculty.washington.edu/browning/intro-to-vcf.html}{VCF} file format.
}
