% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistSig.R
\name{persistSig}
\alias{persistSig}
\title{Mutation signatures in different allele frequency spectrum}
\usage{
persistSig(snv, th_vec_lw, th_vec_up, bsg = NULL)
}
\arguments{
\item{snv}{A dataframe having \emph{sample, chr, pos, ref, alt and/or freq} as it's columns. This dataframe may have more than one sample.}

\item{th_vec_lw}{A vector of lower limits of frequency ranges.}

\item{th_vec_up}{A vector of upper limits of frequency ranges.}
}
\value{
A list of samples having mutational signature corresponding to each allele frequency range
}
\description{
Determine the burden of different mutation signatures acrross different allele frequency ranges.
}
\examples{
load(file = "data/snv_sample.rda")   # load 'snv' dataframe object
mut_sig_per_freq_range=persistSig(snv=snv_sample,th_vec_lw=c(0,0.4), th_vec_up=c(0.1,1))  # list of samples having mutational signature for each allele frequency range i.e., in our example (0.0 - 0.1) & (0.4 - 1.0).

}
\seealso{
\code{\link{vcfToSNV}} to generate \emph{snv} dataframe.
}
